+++ src/logic/systems/object_render.h
@@ -28,7 +28,7 @@ public:
         return "ObjectRender";
     }
 
-    void tick(float dt) override;
+    void tick(double dt) override;
     
 private:
