+++ src/logic/systems/particle_simulation.h
@@ -40,7 +40,7 @@ public:
         return "ParticleSimulation";
     }
 
-    void tick(float dt) override;
+    void tick(double dt) override;
 
 private:
