+++ src/logic/systems/ui.cpp
@@ -27,7 +27,7 @@ UISys::UISys(entt::registry& reg) : System(reg) {
     
 }
 
-void UISys::tick(float dt) {
+void UISys::tick(double dt) {
     
     OPTICK_EVENT();
